bbctl_rca/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pyyaml>=6.0
google-generativeai>=0.8.0
anthropic>=0.40.0
diskcache>=5.6.0
pydantic>=2.0.0
python-dotenv>=1.0.0
